+++ vector/src/main/java/im/vector/app/features/DefaultVectorOverrides.kt
@@ -14,6 +14,7 @@ interface VectorOverrides {
     val forceDialPad: Flow<Boolean>
     val forceLoginFallback: Flow<Boolean>
     val forceHomeserverCapabilities: Flow<HomeserverCapabilitiesOverride>?
+    val forceIdentityChangeBanner: Flow<Boolean>
 }
 
 data class HomeserverCapabilitiesOverride(
@@ -25,4 +26,5 @@ class DefaultVectorOverrides : VectorOverrides {
     override val forceDialPad = flowOf(false)
     override val forceLoginFallback = flowOf(false)
     override val forceHomeserverCapabilities: Flow<HomeserverCapabilitiesOverride>? = null
+    override val forceIdentityChangeBanner = flowOf(false)
 }